ai_agent/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
langchain
chromadb
pypdf
langchain-google-genai
langchain-community